requirements.txt
attrs==23.2.0
Automat==22.10.0
azure-core==1.29.7
azure-identity==1.15.0
azure-storage-blob==12.19.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cssselect==1.2.0
cycler==0.12.1
dnspython==2.5.0
filelock==3.13.1
Flask==3.0.1
Flask-Cors==3.0.10
fonttools==4.47.2
gpxpy==1.6.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isodate==0.6.1
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
lxml==5.1.0
MarkupSafe==2.1.4
matplotlib==3.8.2
msal==1.26.0
msal-extensions==1.1.0
numpy==1.26.3
packaging==23.2
pandas==2.2.0
parsel==1.8.1
pillow==10.2.0
portalocker==2.8.2
Protego==0.3.0
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
PyJWT==2.8.0
pymongo==4.6.1
pyOpenSSL==23.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306;sys_platform == 'win32'
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
scikit-learn==1.4.0
scipy==1.12.0
Scrapy==2.11.0
seaborn==0.13.1
service-identity==23.1.0
setuptools==69.0.3
six==1.16.0
threadpoolctl==3.2.0
tldextract==5.1.1
Twisted==22.10.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
w3lib==2.1.2
Werkzeug==3.0.1
zope.interface==6.1
beautifulsoup4==4.9.3
sqlalchemy==1.3.23
pytest